reduce warnings fix bugs recently added

+++ sakura_core/GrepAgent.h
@@ -177,11 +177,11 @@ private:
 		LONGLONG		nLine,			//	マッチした行番号
 		int				nColumn,		//	マッチした桁番号
 		const wchar_t*	pCompareData,	//	行の文字列
-		int				nLineLen,		//	行の文字列の長さ
+		size_t			nLineLen,		//	行の文字列の長さ
 		size_t			nEolCodeLen,	//	EOLの長さ
 		// マッチした文字列の情報
 		const wchar_t*	pMatchData,		//	マッチした文字列
-		int				nMatchLen,		//	マッチした文字列の長さ
+		size_t			nMatchLen,		//	マッチした文字列の長さ
 		// オプション
 		const GrepOption&	grepOption
 	);
